refactor(web): 重构登录页面相关代码，增加Entry作为通用内容
目的是使登录和注册的共用部分能够共用
