+++ src/web/routes/Entry/Register.tsx
@@ -1,56 +1,17 @@
 import { login } from '@redux/actions/user';
 import { TMemo } from '@shared/components/TMemo';
-import {
-  useTRPGDispatch,
-  useTRPGSelector,
-} from '@shared/hooks/useTRPGSelector';
+import { useTRPGDispatch } from '@shared/hooks/useTRPGSelector';
 import { useTranslation } from '@shared/i18n';
 import { registerAccount } from '@shared/model/player';
-import config from '@shared/project.config';
-import { HiddenInMobile } from '@web/components/HiddenInMobile';
-import Webview from '@web/components/Webview';
 import { checkIsOldApp } from '@web/utils/debug-helper';
 import { handleError } from '@web/utils/error';
-import { Button, Input, Space, Typography } from 'antd';
-import React, { useEffect, useMemo, useState } from 'react';
-import { useHistory } from 'react-router';
+import { Button, Input, Typography } from 'antd';
+import React, { useMemo, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useAsyncFn } from 'react-use';
-import styled from 'styled-components';
-import loginPatternUrl from '../assets/img/login-pattern.svg';
+import { ViewContainer } from './style';
 
-const Root = styled.div`
-  display: flex;
-  height: 100vh;
-  width: 100vw;
-`;
-
-const RegisterContainer = styled(Space).attrs({
-  direction: 'vertical',
-  size: 12,
-})`
-  background-color: ${(props) => props.theme.color.graySet[7]};
-  height: 100vh;
-  width: 568px;
-  padding: 10vh 100px 0;
-  background-image: url(${loginPatternUrl});
-  background-repeat: repeat-y;
-  position: relative;
-
-  ${({ theme }) => theme.mixins.mobile('padding: 10vh 20px 0;')};
-
-  .ant-input,
-  .ant-input-password {
-    background-color: ${(props) => props.theme.color.graySet[7]};
-  }
-`;
-
-const InfoContainer = styled.div`
-  flex: 1;
-  background-color: white;
-`;
-
-const RegisterView: React.FC = TMemo(() => {
+export const RegisterView: React.FC = TMemo(() => {
   const [username, setUsername] = useState('');
   const [password, setPassword] = useState('');
   const [passwordRepeat, setPasswordRepeat] = useState('');
@@ -95,7 +56,7 @@ const RegisterView: React.FC = TMemo(() => {
   }, [username, password, passwordRepeat]);
 
   return (
-    <RegisterContainer>
+    <ViewContainer>
       <h2>{t('注册账号')}</h2>
       <Input
         size="large"
@@ -127,41 +88,11 @@ const RegisterView: React.FC = TMemo(() => {
         {t('注册')}
       </Button>
       <div style={{ textAlign: 'right' }}>
-        <Link to="/login" replace={true}>
+        <Link to="/entry/login" replace={true}>
           <Button type="link">{t('已有账号？现在登录')}</Button>
         </Link>
       </div>
-    </RegisterContainer>
+    </ViewContainer>
   );
 });
 RegisterView.displayName = 'RegisterView';
-
-/**
- * TODO: 需要合并
- */
-function useWatchLoginStatus() {
-  const history = useHistory();
-  const isLogin = useTRPGSelector((state) => state.user.isLogin);
-  useEffect(() => {
-    if (isLogin === true) {
-      history.replace('/main');
-    }
-  }, [isLogin]);
-}
-
-export const Register: React.FC = TMemo(() => {
-  useWatchLoginStatus();
-
-  return (
-    <Root>
-      <RegisterView />
-
-      <HiddenInMobile>
-        <InfoContainer>
-          <Webview src={config.url.loginUrl} />
-        </InfoContainer>
-      </HiddenInMobile>
-    </Root>
-  );
-});
-Register.displayName = 'Register';
