+++ src/web/routes/Entry/style.tsx
@@ -1,0 +1,9 @@
+import { Space } from 'antd';
+import styled from 'styled-components';
+
+export const ViewContainer = styled(Space).attrs({
+  size: 12,
+  direction: 'vertical',
+})`
+  width: 100%;
+`;
