+++ src/web/routes/LoginCheck.tsx
@@ -3,6 +3,8 @@ import { TMemo } from '@shared/components/TMemo';
 import { useTRPGSelector } from '@shared/hooks/useTRPGSelector';
 import { useHistory } from 'react-router';
 
+const loginUrl = '/entry/login';
+
 export const LoginCheck: React.FC = TMemo(() => {
   const isTryLogin = useTRPGSelector((state) => state.user.isTryLogin);
   const isLogin = useTRPGSelector((state) => state.user.isLogin);
@@ -10,8 +12,8 @@ export const LoginCheck: React.FC = TMemo(() => {
 
   useEffect(() => {
     if (isTryLogin === false && isLogin === false) {
-      if (history.location.pathname !== '/login') {
-        history.push('/login');
+      if (history.location.pathname !== loginUrl) {
+        history.push(loginUrl);
       }
     }
   }, [isTryLogin, isLogin, history]);
